feat: Add Detect component
This handles the detection that was previously in the Walk component,
but only for the portion of detecting where you are, namely, if you are
touching floors, stairs, or stairtops.

This is useful because other things (like slices) need to detect if they
are touching those things, but don't walk.