feat: Improve Game Scene to use player's slices
By using the player's slices, we can maintain the progress the player
has made thus far in the level. Since we create the slices every time
the level starts, we must assign them to the player everytime, but we
create them and use the player's slice data if he has it.